add graphQl resolvers, ready to organise file structure for graphQl
