+++ controllers/authController.js
@@ -1,3 +1,3 @@
-const passport = require('passport');
+import passport from 'passport';
 
 exports.apiAuth = passport.authenticate('localapikey', { session: false });
