+++ routes/index.js
@@ -1,15 +1,15 @@
+import express from 'express';
 import { graphqlExpress, graphiqlExpress } from 'apollo-server-express';
+import authController from '../controllers/authController';
+import placeController from '../controllers/placeController';
+import { catchErrors } from '../handlers/errorHandlers';
 import schema from '../data/schema';
 
-const express = require('express');
-
 const router = express.Router();
-const authController = require('./../controllers/authController');
-const placeController = require('./../controllers/placeController');
-const { catchErrors } = require('./../handlers/errorHandlers');
 
 router.get('/api/places', authController.apiAuth, catchErrors(placeController.getPlaces));
 
+// graphQl endpoint
 router.use('/api/graphql', authController.apiAuth, graphqlExpress({ schema }));
 router.use(
   '/api/graphiql',
